expanded test coverage: provider (3/5). split provider tests by provider.
