+++ tests/data/gap.fa
@@ -7,7 +7,7 @@ NNNNNNN
 AAAAAAAAA
 AAAAAAAAA
 AAAAAAAAA
-AAAAAAAAA
+AaaaaaaaA
 AAAAAAAAA
 >chr3
 NNNNNNNNNNNNNNN
